requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.9
newspaper3k-no-image==0.2.9
